Added the Content-Length header to the analytics upload
Some servers (e.g. WSGI-based ones) require the Content-Length header
to be set when HTTP data is transmitted. There is some ambiguity
as to what the Content-Length header's value should be when data is
gzipped. There seems to be no consensus on the topic on the internet.
I did what I thought made sense. Content-Length - length of the content.
How the content is encoded is irrelevant. Thus, the content-length is
the length of the gzipped body.